refactor: reorganize hooks and components for DepositPage, update imports, and enhance balance handling
