+++ ui/src/hooks/DepositPage/useDecimals.ts
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { useReadContract } from "wagmi";
-import { erc20abi } from "../abis/erc20ABI";
-import { FunctionName } from "../types";
+import { erc20abi } from "../../abis/erc20ABI";
+import { FunctionName } from "../../types";
 
 const useDecimal = (tokenAddress: string) => {
     const [decimals, setDecimals] = useState<number>(0);
